Add files via upload
The Climatic changes file gives the data of the climatic condition from the year 1961 to 2022. The data is from Food and Agriculture Organization of the United Nations (FAO). 2022. FAOSTAT Climate Change, Climate Indicators, Temperature change. License: CC BY-NC-SA 3.0 IGO. It gives the change in temperature and the surface temperature.

Based on the data in the file is the SUPERVISED DATA. the values in the table is continuous values of the temperature change in celcius. Climatic change indicator is the DEPENDENT VARIABLE.