Add files via upload
This csv file is the people who travelled in titanic. And number of people who are alive in each family. This table gives the information of passengers list, their tickets, Compartment and etc.

Here the table is the UNSUPERVISED table and all the variables are INDEPENDENT.
The table do not have any continues variables. They are CATEGORY Variables.