Add files via upload
This dataset is about the bitcoin and encompasses a wide range of historical Bitcoin price and volume data. it has open rate, high rate, low rate, closing data and adjustable closing date.

all the variables here are INDEPENDENT. The dataset is the UNSUPERVISED dataset. The values here are categorical data.