Add files via upload
The above dataset is the data of students who got placed from different city i.e Los Angles, Newyork and Chicago. students IQ level and their cpga are given. And the students placement information is given.

Now this dataset is the with continuous and categorical values.
IQ and CGPA are CONTINUES VARIABLES where 
CITY and PLACEMENT is the CATEGORICAL VARIABLES.

The data set is the SUPERVISED DATASET with a DEPENDENT VARIABLE "PLACEMENT" and it purely depends on "CGPA".