Create aeroplane traveller dataset
This dataset gives feedback of the travellers from 8 different airlines. Those eight airlines are the best eight in the world. It has the different feedback for their Seat Comfort, Staff Service, Food & Beverages, Inflight Entertainment,Value For Money and Overall Rating from the passengers view.

This is dataset is the combination of continuous and categorical value.

This is dataset is the SUPERVISED DATASET with dependent variable RECOMMENDED.