Add files via upload
Contains details of cars and their various features. To know the factors affecting prices. So decided to collect data influencing the target. This Dataset contains the car name, year, current prize, their selling prize,kms driven, fuel type and transmission.

This data set is the combination of continuous and categorical data. 
the dataset is the SUPERVISED DATASET with the dependent variable Selling prize which is affected by kms driven.